Extract BaseJdbcAuthenticationConfig
Not every JDBC based connector is using user and password for
authentication. Hence allowing user to set user and password there is
incorrect.
