+++ presto-base-jdbc/src/main/java/io/prestosql/plugin/jdbc/credential/ExtraCredentialProvider.java
@@ -13,7 +13,7 @@
  */
 package io.prestosql.plugin.jdbc.credential;
 
-import io.prestosql.plugin.jdbc.BaseJdbcConfig;
+import io.prestosql.plugin.jdbc.BaseJdbcAuthenticationConfig;
 import io.prestosql.plugin.jdbc.JdbcIdentity;
 
 import javax.inject.Inject;
@@ -31,9 +31,9 @@ public class ExtraCredentialProvider
     private final CredentialProvider delegate;
 
     @Inject
-    public ExtraCredentialProvider(BaseJdbcConfig baseJdbcConfig, @ForExtraCredentialProvider CredentialProvider delegate)
+    public ExtraCredentialProvider(BaseJdbcAuthenticationConfig config, @ForExtraCredentialProvider CredentialProvider delegate)
     {
-        this(Optional.ofNullable(baseJdbcConfig.getUserCredentialName()), Optional.ofNullable(baseJdbcConfig.getPasswordCredentialName()), delegate);
+        this(Optional.ofNullable(config.getUserCredentialName()), Optional.ofNullable(config.getPasswordCredentialName()), delegate);
     }
 
     public ExtraCredentialProvider(Optional<String> userCredentialName, Optional<String> passwordCredentialName, CredentialProvider delegate)
